refactor: reduce main size
