+++ cmd/main.go
@@ -2,17 +2,12 @@ package cmd
 
 import (
 	"fmt"
-	"path/filepath"
-	"regexp"
-	"strings"
 	"sync"
 
 	"github.com/checkmarx/2ms/config"
-	"github.com/checkmarx/2ms/lib"
 	"github.com/checkmarx/2ms/plugins"
 	"github.com/checkmarx/2ms/reporting"
 	"github.com/checkmarx/2ms/secrets"
-	"github.com/rs/zerolog"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -76,34 +71,6 @@ var channels = plugins.Channels{
 var report = reporting.Init()
 var secretsChan = make(chan reporting.Secret)
 
-func initialize() {
-	zerolog.SetGlobalLevel(zerolog.InfoLevel)
-
-	configFilePath, err := rootCmd.Flags().GetString(configFileFlag)
-	if err != nil {
-		cobra.CheckErr(err)
-	}
-	cobra.CheckErr(lib.LoadConfig(vConfig, configFilePath))
-	cobra.CheckErr(lib.BindFlags(rootCmd, vConfig, envPrefix))
-
-	switch strings.ToLower(logLevelVar) {
-	case "trace":
-		zerolog.SetGlobalLevel(zerolog.TraceLevel)
-	case "debug":
-		zerolog.SetGlobalLevel(zerolog.DebugLevel)
-	case "info":
-		zerolog.SetGlobalLevel(zerolog.InfoLevel)
-	case "warn":
-		zerolog.SetGlobalLevel(zerolog.WarnLevel)
-	case "err", "error":
-		zerolog.SetGlobalLevel(zerolog.ErrorLevel)
-	case "fatal":
-		zerolog.SetGlobalLevel(zerolog.FatalLevel)
-	default:
-		zerolog.SetGlobalLevel(zerolog.InfoLevel)
-	}
-}
-
 func Execute() error {
 	vConfig.SetEnvPrefix(envPrefix)
 	vConfig.AutomaticEnv()
@@ -144,23 +111,6 @@ func Execute() error {
 	if report.TotalSecretsFound > 0 {
 		showErrorKind = "results"
 		return fmt.Errorf("")
-	}
-
-	return nil
-}
-
-func validateFormat(stdout string, reportPath []string) error {
-	r := regexp.MustCompile(outputFormatRegexpPattern)
-	if !(r.MatchString(stdout)) {
-		return fmt.Errorf(`invalid output format: %s, available formats are: json, yaml and sarif`, stdout)
-	}
-
-	for _, path := range reportPath {
-		fileExtension := filepath.Ext(path)
-		format := strings.TrimPrefix(fileExtension, ".")
-		if !(r.MatchString(format)) {
-			return fmt.Errorf(`invalid report extension: %s, available extensions are: json, yaml and sarif`, format)
-		}
 	}
 
 	return nil
